+++ Cargo.toml
@@ -22,6 +22,8 @@ rust-version = "1.80"
 [workspace.dependencies]
 anyhow = "1.0.98"
 serde = "1.0.219"
+uniffi = "0.29.1"
+uniffi_bindgen = "0.29.1"
 
 [workspace.lints.rust]
 unsafe_code = "forbid"
