+++ crux_core/Cargo.toml
@@ -36,6 +36,12 @@ thiserror = "2.0.12"
 serde-generate = { version = "=0.26.0", optional = true }
 serde-reflection = { version = "=0.4.0", optional = true }
 
+[target.'cfg(target_family = "wasm")'.dependencies]
+wasm-bindgen = "0.2.100"
+
+[target.'cfg(not(target_family = "wasm"))'.dependencies]
+uniffi.workspace = true
+
 [dev-dependencies]
 assert_fs = "1.1.3"
 assert_matches = "1.5"
