+++ examples/counter-next/shared/.gitignore
@@ -1,2 +1,4 @@
 !/generated
 node_modules
+Shared
+.swiftpm
