+++ examples/counter-next/shared/Cargo.toml
@@ -22,12 +22,14 @@ chrono = { version = "0.4.40", features = ["serde"] }
 crux_core.workspace = true
 crux_http.workspace = true
 futures = "0.3.31"
-lazy_static = "1.5.0"
 serde = { workspace = true, features = ["derive"] }
 serde_json = "1.0.140"
-# uniffi = "0.28.3"
-uniffi = "0.29.1"
 url = "2.5.4"
+
+[target.'cfg(not(target_family = "wasm"))'.dependencies]
+uniffi = "0.29.1"
+
+[target.'cfg(target_family = "wasm")'.dependencies]
 wasm-bindgen = "0.2.100"
 
 [dev-dependencies]
