+++ rust-toolchain.toml
@@ -2,11 +2,11 @@
 channel = "stable"
 components = ["rustfmt", "rustc-dev"]
 targets = [
-  "aarch64-apple-darwin",
-  "aarch64-apple-ios",
-  "aarch64-apple-ios-sim",
-  "aarch64-linux-android",
-  "wasm32-unknown-unknown",
-  "x86_64-apple-ios"
+    "aarch64-apple-darwin",
+    "aarch64-apple-ios",
+    "aarch64-apple-ios-sim",
+    "aarch64-linux-android",
+    "wasm32-unknown-unknown",
+    "x86_64-apple-ios",
 ]
 profile = "minimal"
